+++ examples/blinky_multiple.rs
@@ -18,7 +18,12 @@ fn main() -> ! {
         let gpioa = p.GPIOA.split(&mut rcc);
         let gpiob = p.GPIOB.split(&mut rcc);
 
-        let (led1, led2) = cortex_m::interrupt::free(move |cs| {
+        let (led1, led2) = cortex_m::interrupt::free(move |_| {
+            // SAFETY: We are in a critical section, but the `cortex_m` critical section
+            // token is not compatible with the `bare_metal` token. Future version of the
+            // `cortex_m` crate will not supply *any* token to this callback!
+            let cs = unsafe { &bare_metal::CriticalSection::new() };
+
             (
                 // (Re-)configure PA1 as output
                 gpioa.pa1.into_push_pull_output(cs),