+++ examples/serial_echo.rs
@@ -17,7 +17,12 @@ fn main() -> ! {
 
         let gpioa = p.GPIOA.split(&mut rcc);
 
-        let (tx, rx) = cortex_m::interrupt::free(move |cs| {
+        let (tx, rx) = cortex_m::interrupt::free(move |_| {
+            // SAFETY: We are in a critical section, but the `cortex_m` critical section
+            // token is not compatible with the `bare_metal` token. Future version of the
+            // `cortex_m` crate will not supply *any* token to this callback!
+            let cs = unsafe { &bare_metal::CriticalSection::new() };
+
             (
                 gpioa.pa9.into_alternate_af1(cs),
                 gpioa.pa10.into_alternate_af1(cs),