+++ examples/watchdog.rs
@@ -30,7 +30,13 @@ fn main() -> ! {
         let mut delay = Delay::new(cp.SYST, &rcc);
 
         // Configure serial TX pin
-        let tx = cortex_m::interrupt::free(move |cs| gpioa.pa9.into_alternate_af1(cs));
+        let tx = cortex_m::interrupt::free(move |_| {
+            // SAFETY: We are in a critical section, but the `cortex_m` critical section
+            // token is not compatible with the `bare_metal` token. Future version of the
+            // `cortex_m` crate will not supply *any* token to this callback!
+            let cs = unsafe { &bare_metal::CriticalSection::new() };
+            gpioa.pa9.into_alternate_af1(cs)
+        });
 
         // Obtain a serial peripheral with for unidirectional communication
         let mut serial = Serial::usart1tx(p.USART1, tx, 115_200.bps(), &mut rcc);