+++ src/dac.rs
@@ -29,7 +29,8 @@
 //!#[entry]
 //!fn main() -> ! {
 //!    if let (Some(mut dp), Some(_cp)) = (pac::Peripherals::take(), cortex_m::Peripherals::take()) {
-//!        cortex_m::interrupt::free(move |cs| {
+//!        cortex_m::interrupt::free(move |_| {
+//!            let cs = unsafe { &bare_metal::CriticalSection::new() };
 //!            let mut rcc = dp.RCC.configure().sysclk(8.mhz()).freeze(&mut dp.FLASH);
 //!
 //!            let gpioa = dp.GPIOA.split(&mut rcc);