+++ src/gpio.rs
@@ -190,7 +190,7 @@ macro_rules! gpio {
                     pac::$GPIOX
                 };
 
-                use cortex_m::interrupt::CriticalSection;
+                use bare_metal::CriticalSection;
 
                 use super::{
                     Alternate, Analog, Floating, GpioExt, Input, OpenDrain, Output,